+++ packages/filetypes/lib/register.js
@@ -8,6 +8,7 @@ registry.register({
     types: {
       ceph_conf: '@nikitajs/filetypes/lib/ceph_conf',
       locale_gen: '@nikitajs/filetypes/lib/locale_gen',
+      my_cnf: '@nikitajs/filetypes/lib/my_cnf',
       pacman_conf: '@nikitajs/filetypes/lib/pacman_conf',
       ssh_authorized_keys: '@nikitajs/filetypes/lib/ssh_authorized_keys',
       yum_repo: '@nikitajs/filetypes/lib/yum_repo'